DDS-236 #resolve

+++ app/api/dds/v1/current_user_api.rb
@@ -32,7 +32,8 @@ module DDS
         named 'manage current_user api_key'
         failure [
           [201, "Success"],
-          [401, "Missing, Expired, or Invalid API Token in 'Authorization' Header"]
+          [401, "Missing, Expired, or Invalid API Token in 'Authorization' Header"],
+          [403, 'Forbidden (software_agent restricted)']
         ]
       end
       put '/current_user/api_key', serializer: ApiKeySerializer do
@@ -41,11 +42,13 @@ module DDS
           ApiKey.transaction do
             audits_to_annotate = []
             if current_user.api_key
+              authorize current_user.api_key, :update?
               original_api_key_id = current_user.api_key.id
               current_user.api_key.destroy!
               audits_to_annotate << Audited.audit_class.where(auditable_id: original_api_key_id).last
             end
             current_user.build_api_key(key: SecureRandom.hex)
+            authorize current_user.api_key, :create?
             current_user.save
             audits_to_annotate << current_user.api_key.audits.last
             annotate_audits audits_to_annotate
@@ -60,12 +63,13 @@ module DDS
         failure [
           [200, 'Success'],
           [401, 'Unauthorized'],
-          [403, 'Forbidden'],
+          [403, 'Forbidden (software_agent restricted)'],
           [404, 'Current User Does not Exist']
         ]
       end
       get '/current_user/api_key', serializer: ApiKeySerializer do
         authenticate!
+        authorize current_user.api_key, :show?
         current_user.api_key
       end
 
@@ -75,12 +79,13 @@ module DDS
         failure [
           [200, 'Success'],
           [401, 'Unauthorized'],
-          [403, 'Forbidden'],
+          [403, 'Forbidden (software_agent restricted)'],
           [404, 'Current User Does not Exist']
         ]
       end
       delete '/current_user/api_key', root: false do
         authenticate!
+        authorize current_user.api_key, :destroy?
         Audited.audit_class.as_user(current_user) do
           current_user.api_key.destroy!
           annotate_audits [current_user.api_key.audits.last]
